requirements.txt
Flask>=2.0.0
Pillow>=8.0.0
imagehash>=4.0.0
numpy>=1.19.0
Werkzeug>=2.0.0
